Allow use of just one fog platform
Previously, by requiring the root fog gem to be loaded, all gems
possibly supported by fog would need to be pulled in to the project,
since this is what the 'fog' gem does. Alternately, fog supports just
requiring a specific service, like S3 or Rackspace, by requiring the
services' fog gem specifically.

This commit adds support for the latter use case, allowing developers
to mimimize the number of gems pulled in to a project.