Change default changefreq from always to monthly
Google Search Console will raise an "Invalid tag value" error when a
sitemap entry has a changefreq of "always" but is rarely changed. This
is presumably because Google does not want to be encouraged to crawl a
page unnecessarily, as that is a waste of resources.

This changes the default changefreq from "always" to "monthly" to avoid
causing Google Search Console errors, and to be a better default in
general, because "always" is rarely appropriate and "monthly"
frequently is.

This also unifies the 2 uses of the default changefreq into a single
configuration variable. In the future this configuration variable
could become officially supported.