only set endpoint from legacy option if present
In the `aws-sdk-core` gem, the `endpoint` option is meant for
overriding the default endpoint value resolved from the region
option [1]. Passing a `nil` value for the endpoint option is
not the same as not passing it at all, as nil is still considered
to be an override [2][3].

A recent refactor to the `AwsSdkAdapter` looks to have tweaked the
options logic to always include the `endpoint` key in the object
being passed to `Aws::S3::Resource`. By default, this will raise
a "missing required option :endpoint" exception as the default
legacy option of `nil` will be set for the endpoint.

This change ensures that the `endpoint` value isn't set if
the legacy option is not specified.

[1] https://github.com/aws/aws-sdk-ruby/blob/c97f6932b6d5d6bc3e45aaf1068be52afd6781be/gems/aws-sdk-core/lib/seahorse/client/plugins/endpoint.rb#L11-L15
[2] https://github.com/aws/aws-sdk-ruby/blob/c97f6932b6d5d6bc3e45aaf1068be52afd6781be/gems/aws-sdk-core/lib/seahorse/client/plugins/endpoint.rb#L29-L32
[3] https://github.com/aws/aws-sdk-ruby/issues/2066